chore: tidy reviewer nits

+++ src/mods/foveated/VisibilityCache.cpp
@@ -24,7 +24,7 @@ void VisibilityCache::initialize(uint32_t maxPrimitives) {
 void VisibilityCache::beginFrame(int frameIdx) {
     std::unique_lock lock{m_mtx};
     const auto wrapped = static_cast<uint32_t>(frameIdx);
-    m_current = static_cast<int>(wrapped % FRAMES);
+    m_current = static_cast<int>(wrapped % static_cast<uint32_t>(FRAMES));
     m_frames[m_current].valid.fill(false);
 }
